move faker to bootstrap-application-base
